Change to yaml config
Left the struct tags for toml in the struct definitions, so maybe
we'll add it back later.

We're waiting for the Go toml libraries to update to the latest
TOML spec, which lets us have dot-separated keys so we can move
our map[string][]string off of a single line, which was a bit
oppressive. Yaml gives us an easier option here.